requirements.txt
autopep8==1.4.3
beautifulsoup4==4.7.1
boto==2.49.0
certifi==2018.11.29
chardet==3.0.4
Django==2.0.1
django-admin-sortable==2.1.9
django-appconf==1.0.3
django-background-tasks==1.2.0
django-ckeditor==5.6.1
django-compat==1.0.15
django-htmlmin==0.11.0
django-js-asset==1.1.0
django-post-office==3.1.0
django-ses==0.8.9
django-slack==5.13.0
django-treebeard==4.3
future==0.17.1
geographiclib==1.49
gprof2dot==2016.10.13
html5lib==1.0.1
idna==2.8
instamojo-wrapper==1.1
Jinja2==2.10.1
jsonfield==2.0.2
MarkupSafe==1.1.1
Pillow==5.3.0
psycopg2==2.7.7
psycopg2-binary==2.7.7
pycodestyle==2.5.0
Pygments==2.3.1
python-dateutil==2.8.0
pytz==2018.7
requests==2.21.0
six==1.12.0
sorl-thumbnail==12.5.0
soupsieve==1.9
sqlparse==0.3.0
urllib3==1.24.1
webencodings==0.5.1
